requirements.txt
asgiref
attrs
certifi
cffi
charset-normalizer
configparser
cryptography
csp
dj-stripe
Django
django-raw-api
django-secure
docopt
future
idna
isodate
lxml
Naked
payments
phonenumberslite
pipreqs
platformdirs
pycparser
pylance
pytz
PyYAML
requests-file
requests-toolbelt
shellescape
six
sqlparse
stripe
trafaret
urllib3
yarg
zeep
zipp
django-background-tasks
greenlet
numpy
packaging
pandas
psutil
pyarrow
python-dateutil
ruamel.yaml
sqlalchemy
typing-extensions
tzdata
pycryptodome
django-haystack
django_pdb
charset-normalizer
django-payments
django-phonenumber-field
requests
kavenegar
pillow
elasticsearch>=2.0.0,<3.0.0
gunicorn
